fix: list response plan on get risk and deviation

+++ src/Core/Domain/ResponsePlans/Handlers/ResponsePlansCommandHandler.cs
@@ -1,18 +1,22 @@
 using System.Linq;
 using System.Threading.Tasks;
+using Actions.Core.Domain.Actions.Enums;
 using Actions.Core.Domain.ResponsePlans.Commands;
 using Actions.Core.Domain.ResponsePlans.Dtos;
 using Actions.Core.Domain.ResponsePlans.Entities;
 using Actions.Core.Domain.ResponsePlans.Interfaces;
+using Actions.Core.Domain.Shared.Interfaces.Entities;
 
 namespace Actions.Core.Domain.ResponsePlans.Handlers
 {
     public class ResponsePlansCommandHandler
     {
         private readonly IResponsePlanRepository _repository;
-        public ResponsePlansCommandHandler(IResponsePlanRepository repository)
+        private readonly ITokenUtil _tokenUtil;
+        public ResponsePlansCommandHandler(IResponsePlanRepository repository, ITokenUtil tokenUtil)
         {
             _repository = repository;
+            _tokenUtil = tokenUtil;
         }
 
         public void Handle(ProcessListResponsePlan command)
@@ -26,7 +30,7 @@ namespace Actions.Core.Domain.ResponsePlans.Handlers
 
                 if (process)
                 {
-                    var responsePlansIdDB = responsePlansDB.Select(plan => plan.Id);
+                    var responsePlansIdDB = responsePlansDB != null ? responsePlansDB.Select(plan => plan.Id) : Enumerable.Empty<string>();
 
                     var listToAdd = command.ResponsePlanList
                         .Where(responsePlan => !responsePlansIdDB.Contains(responsePlan.Id))
@@ -34,7 +38,7 @@ namespace Actions.Core.Domain.ResponsePlans.Handlers
                             responsePlan.Description, 
                             responsePlan.Responsible.Id, 
                             responsePlan.DueDate,
-                            responsePlan.Status,
+                            (ActionStatusEnum)responsePlan.Status.Id,
                             responsePlan.ActualStartDate,
                             responsePlan.ActualEndDate,
                             responsePlan.Cost,
@@ -44,25 +48,30 @@ namespace Actions.Core.Domain.ResponsePlans.Handlers
 
                     _repository.InsertRange(listToAdd);
 
-                    var responsePlansIds = command.ResponsePlanList.Select(plan => plan.Id);
-                    var listToUpdate = responsePlansDB
-                        .Where(responsePlan => responsePlansIds.Contains(responsePlan.Id))
-                        .Select(responsePlan => ToResponsePlan(responsePlan));
-
-                    foreach (var item in listToUpdate)
+                    if (responsePlansDB != null)
                     {
-                        var responsePlan = command.ResponsePlanList.SingleOrDefault(plan => plan.Id == item.Id);
+                        var responsePlansIds = command.ResponsePlanList.Select(plan => plan.Id);
+                        var listToUpdate = responsePlansDB
+                            .Where(responsePlan => responsePlansIds.Contains(responsePlan.Id))
+                            .Select(responsePlan => ToResponsePlan(responsePlan));
 
-                        if (item.HasModified(responsePlan.Description, responsePlan.Responsible.Id, responsePlan.DueDate, responsePlan.OriginalDueDate, responsePlan.Status, responsePlan.ActualStartDate, responsePlan.ActualEndDate, responsePlan.Cost, responsePlan.Comments))
-                            _repository.Update(item);
+                        foreach (var item in listToUpdate)
+                        {
+                            var responsePlan = command.ResponsePlanList.SingleOrDefault(plan => plan.Id == item.Id);
+
+                            if (item.HasModified(responsePlan.Description, responsePlan.Responsible.Id, responsePlan.DueDate, responsePlan.OriginalDueDate, (ActionStatusEnum)responsePlan.Status.Id, responsePlan.ActualStartDate, responsePlan.ActualEndDate, responsePlan.Cost, responsePlan.Comments))
+                            {
+                                item.UpdateData(responsePlan.Description, responsePlan.Responsible.Id, responsePlan.DueDate, responsePlan.OriginalDueDate, (ActionStatusEnum)responsePlan.Status.Id, responsePlan.ActualStartDate, responsePlan.ActualEndDate, responsePlan.Cost, responsePlan.Comments, _tokenUtil.Id);
+                                _repository.Update(item);
+                            }
+                        }
+
+                        var listToDelete = responsePlansDB
+                            .Where(responsePlan => !responsePlansIds.Contains(responsePlan.Id))
+                            .Select(responsePlan => ToResponsePlan(responsePlan));
+
+                        _repository.DeleteRange(listToDelete);
                     }
-
-                    var listToDelete = responsePlansDB
-                        .Where(responsePlan => !responsePlansIds.Contains(responsePlan.Id))
-                        .Select(responsePlan => ToResponsePlan(responsePlan));
-
-                    _repository.DeleteRange(listToDelete);
-
                 }
             }
         }
@@ -75,7 +84,7 @@ namespace Actions.Core.Domain.ResponsePlans.Handlers
                 responsePlan.Responsible.Id, 
                 responsePlan.DueDate,
                 responsePlan.OriginalDueDate,
-                responsePlan.Status,
+                (ActionStatusEnum)responsePlan.Status.Id,
                 responsePlan.ActualStartDate,
                 responsePlan.ActualEndDate,
                 responsePlan.Cost,
